requirements.txt
sqlalchemy==1.4.18
python-dotenv==0.14.0
beautifulsoup4==4.9.1
discord.py==1.7.3
git+https://github.com/Rapptz/discord-ext-menus@master
discord-ext-ipc==2.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.6.1
pytz==2019.3
requests==2.24.0
requests-unixsocket==0.1.5
tabulate==0.8.7
yarl==1.4.2
feedparser==6.0.2
googletrans==4.0.0rc1
quart==0.15.1
Quart-CORS==0.5.0
attrs~=21.2.0
dacite~=1.6.0
pytest==6.2.4
markdownify==0.9.2

# Experimental package for slash commands & menus
dislash.py==1.4.9